refact: 테스트 파일과 프로토콜 정의 분리
